+++ code/frontend/src/app/components/shared/ingredient-card/ingredient-card.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { IngredientDto } from 'src/app/dtos';
 import { IngredientUnit } from 'src/app/enums';
 import { IngredientService } from 'src/app/services';
+import { ConfirmationBoxService } from 'src/app/services/confirmation-box.service';
 
 @Component({
   selector: 'app-ingredient-card',
@@ -14,6 +15,7 @@ export class IngredientCardComponent {
   constructor(
     private router: Router,
     private ingredientService: IngredientService,
+    private confirmationBoxService: ConfirmationBoxService
   ) { }
 
   @Input() ingredient: IngredientDto;
@@ -42,12 +44,17 @@ export class IngredientCardComponent {
   }
 
   deleteIngredient(): void {
-    this.ingredientService.delete(this.ingredient.id).subscribe(
-      (data) => {
-        console.log(data);
-        this.deletedElement.emit();
-      }
-    );
+    this.confirmationBoxService.confirm('Delete ingredient', 'Are you sure you want to delete this ingredient?').then(
+      (confirmed) => {
+        if (confirmed) {
+          this.ingredientService.delete(this.ingredient.id).subscribe(
+            (data) => {
+              console.log(data);
+              this.deletedElement.emit();
+            }
+          );
+        }
+    });
   }
 
   goToDetailedView(): void {
